Added reset password modal to account page
